+++ examples/example_canvas.rs
@@ -35,13 +35,13 @@ impl actix::Handler<woab::Signal> for WindowActor {
                 );
                 draw_ctx.set_source_rgb(0.5, 0.5, 0.5);
                 draw_ctx.fill().unwrap();
-                Some(gtk::Inhibit(false))
+                Some(glib::Propagation::Stop)
             }
             "configure_draw_area" => {
                 let event: gdk::EventConfigure = msg.event_param()?;
                 let (width, height) = event.size();
                 self.area_size = [width as f64, height as f64];
-                Some(gtk::Inhibit(false))
+                Some(glib::Propagation::Stop)
             }
             _ => msg.cant_handle()?,
         })