+++ src/lib.rs
@@ -21,7 +21,7 @@
 //! actor which will receive the signals defined in the Glade GTK or connect them to an existing
 //! actor and tag the signals (so that multiple instances can be added - e.g. with `GtkListBox` -
 //! and the signal handler can know from which one the event came). The actors receive the signals
-//! as Actix messages, and the `Handler` returns the inhibitness decision (if the signal requires
+//! as Actix messages, and the `Handler` returns the propagation decision (if the signal requires
 //! it)
 //!
 //! To remove widget-bound actors at runtime, see [`woab::Remove`](Remove).
@@ -73,7 +73,7 @@
 //!             "sig2" => {
 //!                 let woab::params!(text_buffer: gtk::TextBuffer, _) = msg.params()?;
 //!                 // Behavior for sig2 that uses the signal parameters.
-//!                 Some(gtk::Inhibit(false)) // GTK expects sig2 to return its inhibitness decision
+//!                 Some(glib::Propagation::Stop) // GTK expects sig2 to return its propagation decision
 //!             },
 //!             _ => msg.cant_handle()?,
 //!         })